requirements.txt
numpy
transformers
torch
biopython
qdrant-client
